requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
ddgs>=9.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-docx>=1.1.0
Pillow>=10.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
